refactor: split out intent system into intent.rs
